+++ src/machinable/execution/execution.py
@@ -42,6 +42,19 @@ def to_color(experiment_id):
     )
 
 
+def _code_backup_settings(value):
+    enabled = None
+    exclude = None
+    if value in [True, False, None]:
+        enabled = value
+    if isinstance(value, dict):
+        enabled = value.get("enabled", None)
+        exclude = value.get("exclude", None)
+    if isinstance(exclude, tuple):
+        exclude = list(exclude)
+    return {"enabled": enabled, "exclude": exclude}
+
+
 _latest = [None]
 
 
@@ -373,12 +386,33 @@ class Execution(Jsonable):
             self.engine.on_before_storage_creation(self)
             Registration.get().on_before_storage_creation(self)
 
+            # determine code backup settings
+            code_backup = _code_backup_settings(self.code_backup)
+
+            # use project settings by default
+            _project_settings = _code_backup_settings(
+                Registration.get().default_code_backup(execution=self)
+            )
+
+            if code_backup["enabled"] is None:
+                code_backup["enabled"] = _project_settings["enabled"]
+            if code_backup["exclude"] is None:
+                code_backup["exclude"] = _project_settings["exclude"]
+
+            # otherwise fall back on system-wide settings
+            _user_settings = _code_backup_settings(
+                get_settings()["default_code_backup"]
+            )
+            if code_backup["enabled"] is None:
+                _user_settings["enabled"] = _user_settings["enabled"]
+            if code_backup["exclude"] is None:
+                code_backup["exclude"] = _user_settings["exclude"]
+
             # do not backup on mem:// filesystem unless explicitly set to True
-            code_backup = self.code_backup
-            if code_backup is None and not self.storage.config["url"].startswith(
-                "mem://"
-            ):
-                code_backup = True
+            if code_backup["enabled"] is None and not self.storage.config[
+                "url"
+            ].startswith("mem://"):
+                code_backup["enabled"] = True
 
             # collect and write execution data
             url = self.storage.get_url()
@@ -403,8 +437,10 @@ class Execution(Jsonable):
             }
 
             with open_fs({"url": url, "create": True}) as filesystem:
-                if code_backup:
-                    self.project.backup_source_code(opener=filesystem.open)
+                if code_backup["enabled"]:
+                    self.project.backup_source_code(
+                        opener=filesystem.open, exclude=code_backup["exclude"]
+                    )
 
                 for k, v in data.items():
                     filesystem.save_file(name=k, data=v)